For #21035 - Refactor HistoryViewInteractor from HistoryView into HistoryInteractor
